Add a simple Ternus display demo.
Add both an HTML file to create the display and a JS file to control the
animation.  Displays two circles flashing back and forth with a fixed
duration and adjustable ISI.  The ISI adjustment is fairly primitive at
this point: the user can cycle through a set of ISIs by clicking anywhere
on the animation display.

This demo may not be very useful: the type of animation used cannot handle
ISIs shorter than about 25 ms, and doesn't make it convincing that the dots
ever disappeared at that short ISI.  As best as I can understand things,
this the best JavaScript can do.